Add a status telstate attribute for cam2telstate
Some of the code is shuffled around a bit to make the logic a bit
simpler and reduce the nesting of try blocks. The crux is that we track
which sensors still haven't had an initial value set (where a status
like 'unknown' or 'unreachable' counts as an initial value, even though
nothing is entered into telstate).

This also showed up a bug: pool_resources is subscribed both in the
initialisation (to find out which sensors to use) and in the bulk sensor
list, but the updates sent for it in the latter case get swallowed. This
is now fixed.