Support split CBF and SDP proxies
The latest CAM incarnation splits the data proxy into separate CBF and
SDP proxies. Split the sensors accordingly. We look for the old data
proxy in the pool resources as it is still supported. The SP name for
both proxies is also still 'data' for katdal compatibility. The CMC
client has now shifted from the MCP to the CBF proxy. Its sensors
might not work with old-style data proxies, though.