Fix instrument names in sensors
The new CAM setup has a proxy_instrument_stream_xxx format for sensor
names. The CBF device under the proxy therefore has the instrument name
instead of 'cbf'. Also rename _stream_types to be more descriptive
(it is primarily a list of streams, with associated types).

It is still unclear whether this will work in the old 'data' regime.